+++ src/store/zenledger/zenledger.models.ts
@@ -1,4 +1,4 @@
-export type ZLRequestWalletsType = {
+export type ZenLedgerRequestWalletsType = {
   address: string;
   blockchain: string;
   display_name: string;
